hex_defconfig: Heavily adjust boosting tunables
The goal here is to potentially reduce (or remove) direct boosting
entirely and rely purely on Dynamic SchedTune boosting. This should
significantly benefit battery life.

Wake boost was disabled because it is not efficient enough to warrant
the power cost of maxing the cpus for over a second.

General boosting can stay for the case that a jump to a higher frequency
will arise soon. We can prevent any possible stutter if we can keep the
minimum frequency near 600mhz for a few seconds after the user releases
their finger from the screen.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>